+++ Basil/include/Scene/Component.h
@@ -23,20 +23,15 @@ namespace Basil
 	struct NativeScriptComponent
 	{
 		ScriptableEntity* instance = nullptr;
-		std::function<void()> instantiateFunction;
-		std::function<void()> destroyInstanceFunction;
-		std::function<void(ScriptableEntity*)> onCreateFunction;
-		std::function<void(ScriptableEntity*)> onDestroyFunction;
-		std::function<void(ScriptableEntity*, Timestep)> onUpdateFunction;
+
+		ScriptableEntity*(*instantiateScript)();
+		void (*destoryScript)(NativeScriptComponent*);
 
 		template <typename T>
 		void bind()
 		{
-			instantiateFunction = [&]() { instance = new T(); };
-			destroyInstanceFunction = [&]() { delete (T*)instance; instance = nullptr; };
-			onCreateFunction = [](ScriptableEntity* instance) { ((T*)instance)->onCreate(); };
-			onDestroyFunction = [](ScriptableEntity* instance) { ((T*)instance)->onDestroy(); };
-			onUpdateFunction = [](ScriptableEntity* instance, Timestep timeStep) { ((T*)instance)->onUpdate(timeStep); };
+			instantiateScript = []() { return static_cast<ScriptableEntity*>(new T()); };
+			destroyScript = [](NativeScriptComponent* nsc) { delete nsc->instance; nsc->instance = nullptr; }
 		}
 	};