+++ Basil/include/Scene/ScriptableEntity.h
@@ -10,11 +10,17 @@ namespace Basil
 	class ScriptableEntity
 	{
 		public:
+			virtual ~ScriptableEntity() {}
+
 			template <typename T>
 			T& getComponent()
 			{
 				return entity.getComponent<T>();
 			}
+		protected:
+			virtual void onCreate() {}
+			virtual void onDestroy() {}
+			virtual void onUpdate(Timestep timeStep) {}
 		private:
 			Entity entity;
 			friend class Scene;